# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Country": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"communities": {
						"items": {
							"$ref": "#/components/schemas/Community"
						},
						"type": "array"
					},
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/Address"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"communities",
					"addresses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Community": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"provinces": {
						"items": {
							"$ref": "#/components/schemas/Province"
						},
						"type": "array"
					},
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/Address"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"country",
					"provinces",
					"addresses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Province": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"community": {
						"$ref": "#/components/schemas/Community"
					},
					"cities": {
						"$ref": "#/components/schemas/City"
					},
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/Address"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"community",
					"cities",
					"addresses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Address": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"community": {
						"$ref": "#/components/schemas/Community"
					},
					"province": {
						"$ref": "#/components/schemas/Province"
					},
					"city": {
						"$ref": "#/components/schemas/City"
					},
					"street_name": {
						"type": "string"
					},
					"street_number": {
						"type": "string"
					},
					"floor": {
						"type": "number",
						"format": "double"
					},
					"door": {
						"type": "string"
					},
					"zip_code": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"is_active": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"country",
					"community",
					"province",
					"city",
					"street_name",
					"street_number",
					"floor",
					"door",
					"zip_code",
					"user",
					"is_active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Timestamp": {
				"type": "string",
				"description": "Timestamp."
			},
			"City": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"province": {
						"$ref": "#/components/schemas/Province"
					},
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/Address"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"province",
					"addresses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"personal_email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"password_token": {
						"type": "string"
					},
					"birth_date": {
						"type": "string",
						"format": "date-time"
					},
					"gender": {
						"type": "number",
						"format": "double"
					},
					"mobile_phone": {
						"type": "string"
					},
					"home_phone": {
						"type": "string"
					},
					"user_image": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"is_active": {
						"type": "number",
						"format": "double"
					},
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/Address"
						},
						"type": "array"
					},
					"groups": {
						"items": {
							"$ref": "#/components/schemas/Group"
						},
						"type": "array"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array"
					},
					"job_area": {
						"$ref": "#/components/schemas/JobArea"
					},
					"position": {
						"$ref": "#/components/schemas/Position"
					},
					"languages": {
						"items": {
							"$ref": "#/components/schemas/Language"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"last_name",
					"username",
					"email",
					"personal_email",
					"password",
					"password_token",
					"birth_date",
					"gender",
					"mobile_phone",
					"home_phone",
					"user_image",
					"bio",
					"is_active",
					"addresses",
					"groups",
					"permissions",
					"job_area",
					"position",
					"languages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Permission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"parent": {
						"$ref": "#/components/schemas/Permission"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array"
					},
					"groups": {
						"items": {
							"$ref": "#/components/schemas/Group"
						},
						"type": "array"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"parent",
					"children",
					"groups",
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Group": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"is_active": {
						"type": "number",
						"format": "double"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"is_active",
					"users",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JobArea": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"is_active": {
						"type": "number",
						"format": "double"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"positions": {
						"items": {
							"$ref": "#/components/schemas/Position"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"is_active",
					"users",
					"positions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Position": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"is_active": {
						"type": "number",
						"format": "double"
					},
					"job_area": {
						"$ref": "#/components/schemas/JobArea"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"is_active",
					"job_area",
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Language": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"updated_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"name": {
						"type": "string"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"is_active": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"deleted_at",
					"name",
					"users",
					"is_active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteResult": {
				"description": "Result object returned by DeleteQueryBuilder execution.",
				"properties": {
					"raw": {
						"description": "Raw SQL result returned by executed query."
					},
					"affected": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Number of affected rows/documents\nNot all drivers support this"
					}
				},
				"required": [
					"raw"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "node_ts_api_rest",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/address/store": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Address"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new address into the database",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Address attributes",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Address",
								"description": "Address attributes"
							}
						}
					}
				}
			}
		},
		"/address/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Address"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get an existing address",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"description": "Address id to get",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Address"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Update an existing address",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"description": "Address id to update",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "New Address attributes",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Address",
								"description": "New Address attributes"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResult"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete an existing address",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"description": "Address id to delete",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/country/store": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Country"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new Country",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Country"
							}
						}
					}
				}
			}
		},
		"/country": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Country"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get one Country",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Country"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": {
											"error": false,
											"message": "Ok",
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all country",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": []
			}
		},
		"/country/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Country"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Update one Country",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Country"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResult"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete one Country",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/province/store": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Province"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new province",
				"tags": [
					"Provinces"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Province"
							}
						}
					}
				}
			}
		},
		"/province/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Province"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get one province",
				"tags": [
					"Provinces"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Province"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Update a province",
				"tags": [
					"Provinces"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Province"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResult"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete a province",
				"tags": [
					"Provinces"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/group/store": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Group"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new Group",
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Group"
							}
						}
					}
				}
			}
		},
		"/group": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Group"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get one Group",
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Group"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": {
											"error": false,
											"message": "Ok",
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all group",
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": []
			}
		},
		"/group/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Group"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Update one Group",
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Group"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResult"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Ok",
											"error": false,
											"code": 201,
											"result": {}
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete one Group",
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}